+++ src/lib.rs
@@ -19,11 +19,11 @@
 //!     let layer = tile.create_layer("First Layer");
 //!     // NOTE: normally, the Transform would come from MapGrid::tile_transform
 //!     let b = GeomEncoder::new(GeomType::Linestring, Transform::default())
-//!         .point(0.0, 0.0)
-//!         .point(1024.0, 0.0)
-//!         .point(1024.0, 2048.0)
-//!         .point(2048.0, 2048.0)
-//!         .point(2048.0, 4096.0)
+//!         .point(0.0, 0.0)?
+//!         .point(1024.0, 0.0)?
+//!         .point(1024.0, 2048.0)?
+//!         .point(2048.0, 2048.0)?
+//!         .point(2048.0, 4096.0)?
 //!         .encode()?;
 //!     let mut feature = layer.into_feature(b);
 //!     feature.set_id(1);