+++ src/tile.rs
@@ -76,8 +76,8 @@ pub struct Layer {
 /// let tile = Tile::new(4096);
 /// let layer = tile.create_layer("First Layer");
 /// let geom_data = GeomEncoder::new(GeomType::Point, Transform::default())
-///     .point(1.0, 2.0)
-///     .point(7.0, 6.0)
+///     .point(1.0, 2.0)?
+///     .point(7.0, 6.0)?
 ///     .encode()?;
 /// let feature = layer.into_feature(geom_data);
 /// // ...